change: test demoqa.com Checkbox move to UndergogsGroupTest, delete old class DemoQACheckboxTest
