+++ src/test/java/school/redrover/UnderdogsGroupTest.java
@@ -94,6 +94,36 @@ public class UnderdogsGroupTest extends BaseTest {
         }
     }
 
+    @Test
+    public void testHomeCheckbox() {
+        getDriver().manage().timeouts().implicitlyWait(Duration.ofMillis(500));
+        getDriver().get("https://demoqa.com/checkbox");
+        getDriver().findElement(new By.ByCssSelector("[class='rct-icon rct-icon-uncheck']")).click();
+
+        WebElement results = getDriver().findElement(By.id("result"));
+        String textHome = """
+                You have selected :
+                home
+                desktop
+                notes
+                commands
+                documents
+                workspace
+                react
+                angular
+                veu
+                office
+                public
+                private
+                classified
+                general
+                downloads
+                wordFile
+                excelFile""";
+
+        Assert.assertEquals(results.getText(), textHome);
+    }
+
     @Ignore
     @Test
     public void numberOfTheCarsPresented() {
